gpt_notebook/cloud_function/requirements.txt
functions-framework==3.*
pandas-gbq==0.19.1
google-generativeai==0.7.0